refactor: moved Modals to separate components
