+++ debarshi-memory-game/src/App.js
@@ -1,26 +1,16 @@
-import { useEffect, useRef, useState } from 'react'
+import { useEffect, useState } from 'react'
 import { Toaster } from 'react-hot-toast'
-import { Box, Button, Grid, TextField, Typography } from '@mui/material'
+import { Box, Button, Grid, Typography } from '@mui/material'
 import { useCardContext } from 'contexts/CardContext'
-import { Card, Modal, Timer } from 'components'
+import { Card, FormModal, GameOverModal, Timer } from 'components'
 import { modals } from 'constants/modals'
 import { useModalContext } from 'contexts/ModalContext'
 import levels from 'constants/levels'
 
-const boxStyles = {
-  backgroundColor: '#fff',
-  width: '20rem',
-  height: '12rem',
-  padding: '1rem',
-  margin: '0 auto',
-  borderRadius: '4px',
-}
-
 const App = () => {
   const [username, setUsername] = useState('')
-  const inputRef = useRef()
 
-  const { cardList, selected, points, turns, reset, onCardSelected, onLevelChange } =
+  const { cardList, selectedCards, points, turns, reset, onCardSelected, onLevelChange } =
     useCardContext()
   const { state, dispatch } = useModalContext()
 
@@ -32,53 +22,16 @@ const App = () => {
 
   const handleOnCardClick = (item) => onCardSelected(item)
 
-  const handleModal = (modalType) => dispatch({ type: modalType })
-
-  const handleStart = () => {
-    if (inputRef.current.value !== '') {
-      setUsername(inputRef.current.value)
-      handleModal(null)
-      inputRef.current.value = ''
-    }
+  const handleStart = (value) => {
+    setUsername(value)
   }
 
   const resetGame = () => {
     setUsername('')
-    handleModal(modals.START_MODAL)
     reset()
   }
 
   const handleOnLevelChange = (current) => onLevelChange(current)
-
-  const renderFormModal = () => (
-    <Modal open={state.modal === modals.START_MODAL} onClose={handleModal}>
-      <Box sx={boxStyles}>
-        <Typography component="p" variant="p" align="center" sx={{ marginBottom: '1rem' }}>
-          Enter your name to start the game
-        </Typography>
-        <TextField label="Enter your name" variant="outlined" fullWidth inputRef={inputRef} />
-        <Button variant="outlined" fullWidth sx={{ marginTop: '1rem' }} onClick={handleStart}>
-          Start
-        </Button>
-      </Box>
-    </Modal>
-  )
-
-  const renderGameOverModal = () => (
-    <Modal open={state.modal === modals.END_MODAL} onClose={() => handleModal(modals.END_MODAL)}>
-      <Box sx={boxStyles}>
-        <Typography component="h4" variant="h4" align="center" sx={{ marginBottom: '1rem' }}>
-          GAME OVER
-        </Typography>
-        <Typography component="p" variant="p" align="center" sx={{ marginBottom: '1rem' }}>
-          Total points : {points}
-        </Typography>
-        <Button variant="contained" fullWidth sx={{ marginTop: '1rem' }} onClick={resetGame}>
-          Reset
-        </Button>
-      </Box>
-    </Modal>
-  )
 
   return (
     <div className="app">
@@ -109,8 +62,8 @@ const App = () => {
         </Typography>
       </Box>
 
-      {renderFormModal()}
-      {renderGameOverModal()}
+      <FormModal onStart={handleStart} />
+      <GameOverModal onReset={resetGame} />
 
       <Grid container columns={{ xs: 4, sm: 8, md: 12 }} spacing={2} sx={{ marginBottom: '2rem' }}>
         {cardList.map((card) => (
@@ -118,7 +71,7 @@ const App = () => {
             <Card
               card={card}
               initialFlip={!state.modal}
-              flip={card.id === selected?.first?.id || card.id === selected?.second?.id}
+              flip={card.id === selectedCards?.first?.id || card.id === selectedCards?.second?.id}
               onClick={handleOnCardClick}
             />
           </Grid>
