+++ debarshi-memory-game/src/components/index.js
@@ -1,3 +1,5 @@
 export { default as Card } from './Card/Card'
 export { default as Modal } from './Modal/Modal'
 export { default as Timer } from './Timer/Timer'
+export { default as FormModal } from './Modal/FormModal'
+export { default as GameOverModal } from './Modal/GameOverModal'
