+++ debarshi-memory-game/src/styles/styles.js
@@ -1,0 +1,10 @@
+const boxStyles = {
+  backgroundColor: '#fff',
+  width: '20rem',
+  height: '12rem',
+  padding: '1rem',
+  margin: '0 auto',
+  borderRadius: '4px',
+}
+
+export { boxStyles }
